requirements.txt
django
djangorestframework
django_extensions
djangorestframework-simplejwt
